Create private routes, refactoring callback page and login backend
The frontend routes was modified to become private for authenticated users.
The modification of the callback page and the modification of the authentication were started.
Now when a user login, is checked and registered on the backend.
If the user don't has a valid login will be create one and save it and finally the user was be sended to frontend.